requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pygame==2.5.2
pycryptodome==3.19.0
pillow==10.0.1
